file exist tests.

+++ bin/create_links.sh
@@ -17,13 +17,22 @@ fi
 
 
 function ln_boot() {
-    echo creating symlink to boot/isolinux/$1 in tftpboot root
-    ln -s boot/isolinux/$1 ./$1
+    if [ -f $1 ]; then
+        echo the symlink $1 already exist
+    else
+        echo creating symlink to boot/isolinux/$1 in tftpboot root
+        ln -s boot/isolinux/$1 ./$1
+    fi
+
 }
 
 function ln_pxe() {
-    echo creating symlink from pxelinux.cfg/default to $1
-    ln -s default ./$1
+    if [ -f $1 ]; then
+        echo the symlink $1 already exist
+    else
+        echo creating symlink from pxelinux.cfg/default to $1
+        ln -s default ./$1
+    fi
 }
 
 # create root links to .c32 files.
